query module hierarchy
