+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition/Function.hs
@@ -57,6 +57,7 @@ import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.Type.List
 import Squeal.PostgreSQL.Query
+import Squeal.PostgreSQL.Query.Values
 import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Type.Schema
